extracted transmitter into its own class. wrote test for transmitter.
